# AppKube-common-security-rbac-services/Api-Specifications/apispecs.yml
openapi: 3.0.3
info:
  title: Common security rbac services
  version: 1.0.0
  description: API for managing organizations, policies, and emails
servers:
  - url: http://localhost:8080/api/v1
...
paths:
...
  /permission-categories:
    get:
      tags:
        - Permissioncategory
      summary: Get all permission categories
      responses:
       '200':
         description: OK
         content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/PermissionCategory'
                
    post:
      tags:
        - Permissioncategory
      summary: Create a new permission category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PermissionCategory'
      responses:
        '200':
          description: Successful.
  /permission-categories/{id}:
    get:
      tags:
        - Permissioncategory
      summary: Get a permission category by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermissionCategory"
        '404':
          description: Not Found
    put:
      tags:
        - Permissioncategory
      summary: Update a permission category
    
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermissionCategory"
        '404':
          description: Not Found
    delete:
      tags:
        - Permissioncategory
      summary: Delete a permission category by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation.
         
        '404':
          description: Not Found
...
components:
  schemas:
    Organization:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        fax:
          type: string
        dateOfEstablishment:
          type: string
          format: date-time
        status:
          type: string
        cmdbOrgId:
          type: integer
          format: int64
        fileName:
          type: string
        fileStorageLocationType:
          type: string
        profileImage:
          type: string
...
    Permission:
      properties:
        id:
          type: integer    
        version:
          type: integer
        name:
          type: string
        description:
          type: string
        status:
          type: string
        organization:
          type: object
          $ref: "#/components/schemas/Organization"
...
    PermissionCategory:
     type: object
     properties:
        id:
         type: integer
         format: int64
        name:
         type: string
        description:
         type: string
        version:
         type: integer
         format: int64
        status:
         type: string
        organization:
         $ref: '#/components/schemas/Organization'
        permissions:
         type: array
         items:
          $ref: '#/components/schemas/Permission'
...